requirements/base.txt
Flask==1.0.2
giphy-client==1.0.0
gunicorn==19.9.0
python-dotenv==0.9.1
